pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-agent-system"
version = "0.1.0"
description = "RAG agent system with MCP tools and local models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.2.0",
    "watchdog>=3.0.0",
    "ollama>=0.2.0",
    "textract>=1.6.0",
    "pypdf>=3.0.0",
    "python-docx>=1.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "aiofiles>=23.0.0",
    "httpx>=0.25.0",
    "mcp>=1.0.0",
    "python-magic>=0.4.27",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.0.280",
]

[project.scripts]
rag-agent = "src.main:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "T20"]
ignore = ["E501", "S101"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]
exclude = ["tests*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
